+++ src/components/super/i-block/modules/activation/index.ts
@@ -40,8 +40,8 @@ const
  * Activates the component.
  * A deactivated component won't load data from providers on initializing.
  *
- * Basically, you don't need to think about component activation,
- * because it automatically synchronizes with the `keep-alive` mode or a special component prop.
+ * Essentially, you don't need to worry about component activation,
+ * as it automatically synchronizes with the `keep-alive` mode or a specific component prop.
  *
  * @param component
  * @param [force] - if true, then the component will be forced to be activated, even if it is already activated
@@ -117,14 +117,13 @@ export function activate(component: iBlock, force?: boolean): void {
  * Deactivates the component.
  * A deactivated component won't load data from providers on initializing.
  *
- * Basically, you don't need to think about component activation,
- * because it automatically synchronizes with the `keep-alive` mode or a special component prop.
+ * Essentially, you don't need to worry about component activation,
+ * as it automatically synchronizes with the `keep-alive` mode or a specific component prop.
  *
  * @param component
  */
 export function deactivate(component: iBlock): void {
-	const
-		{unsafe} = component;
+	const {unsafe} = component;
 
 	if (unsafe.lfc.isBeforeCreate()) {
 		return;
@@ -141,10 +140,11 @@ export function deactivate(component: iBlock): void {
 			callMethodFromComponent(component, 'deactivated');
 		}).catch(stderr);
 
-		// It's important to deactivate the component ASAP to prevent any unexpected rerenders
-		// because state of the component might change during the deactivation process
-		// but not before calling runHook because otherwise the onHookChange event listeners will be muted
-		// and child dynamic components won't receive the deactivation signal
+		// It's important to deactivate the component ASAP to prevent any unexpected re-renders.
+		// The state of the component might change during the deactivation process,
+		// but it is crucial to call runHook before deactivation.
+		// This ensures that the onHookChange event listeners are not muted
+		// and that child dynamic components receive the deactivation signal.
 		onDeactivated(component);
 	}
 }
@@ -155,9 +155,8 @@ export function deactivate(component: iBlock): void {
  * @param component
  * @param [force] - if true, then the component will be forced to be activated, even if it is already activated
  */
-export function onActivated(component: iBlock, force?: boolean): void {
-	const
-		{unsafe} = component;
+export function onActivated(component: iBlock, force: boolean = false): void {
+	const {unsafe} = component;
 
 	const cantActivate =
 		unsafe.isActivated ||
@@ -213,8 +212,7 @@ export function onActivated(component: iBlock, force?: boolean): void {
  * @param component
  */
 export function onDeactivated(component: iBlock): void {
-	const
-		{unsafe} = component;
+	const {unsafe} = component;
 
 	const async = [
 		unsafe.$async,
@@ -227,8 +225,7 @@ export function onDeactivated(component: iBlock): void {
 				return;
 			}
 
-			const
-				fn = $a[`mute-${asyncNames[key]}`.camelize(false)];
+			const fn = $a[`mute-${asyncNames[key]}`.camelize(false)];
 
 			if (Object.isFunction(fn)) {
 				fn.call($a);