+++ src/components/super/i-block/test/unit/deactivation.ts
@@ -11,20 +11,19 @@ import type { JSHandle } from 'playwright';
 import test from 'tests/config/unit/test';
 import { Component } from 'tests/helpers';
 
-import type bSuperIBlockDeactivationDummy from 'components/super/i-block/test/b-super-i-block-deactivation-dummy/b-super-i-block-deactivation-dummy';
 import type { ComponentElement } from 'core/component';
+import type bSuperIBlockDeactivationDummy from 'components/super/i-block/test/b-super-i-block-deactivation-dummy/b-super-i-block-deactivation-dummy';
 
 test.describe('<i-block> component deactivation', () => {
-	let
-		target: JSHandle<bSuperIBlockDeactivationDummy>;
+	let target: JSHandle<bSuperIBlockDeactivationDummy>;
 
 	test.beforeEach(async ({demoPage, page}) => {
 		await demoPage.goto();
 		target = await Component.createComponent<bSuperIBlockDeactivationDummy>(page, 'b-super-i-block-deactivation-dummy');
 	});
 
-	test.describe('when parent component is deactivated', () => {
-		test('children button should be deactivated', async () => {
+	test.describe('when the parent component is deactivated', () => {
+		test('children should also be deactivated', async () => {
 			const state = await target.evaluate((ctx) => {
 				ctx.deactivate();
 				return ctx.unsafe.$refs.button1.hook;
@@ -33,7 +32,7 @@ test.describe('<i-block> component deactivation', () => {
 			test.expect(state).toBe('deactivated');
 		});
 
-		test('button on teleported component should be deactivated', async () => {
+		test('children of a teleported component should be deactivated', async () => {
 			const state = await target.evaluate((ctx) => {
 				ctx.deactivate();
 				return ctx.unsafe.$refs.button2.hook;
@@ -42,8 +41,7 @@ test.describe('<i-block> component deactivation', () => {
 			test.expect(state).toBe('deactivated');
 		});
 
-		test('dynamic child of component should be deactivated', async () => {
-
+		test('dynamic children of a component should be deactivated', async () => {
 			const result = await target.evaluate((ctx) => {
 				const vnode = ctx.vdom.create('b-radio-button');