+++ src/components/super/i-block/i-block.ts
@@ -11,7 +11,7 @@
  * @packageDocumentation
  */
 
-import { component, UnsafeGetter } from 'core/component';
+import { component, hook, watch, UnsafeGetter } from 'core/component';
 import type { Classes } from 'components/friends/provide';
 
 import type { ModVal, ModsDecl, ModsProp, ModsDict } from 'components/super/i-block/modules/mods';
@@ -25,6 +25,8 @@ import('components/super/i-block/test/b-super-i-block-mods-dummy');
 import('components/super/i-block/test/b-super-i-block-watch-dummy');
 import('components/super/i-block/test/b-super-i-block-lfc-dummy');
 import('components/super/i-block/test/b-super-i-block-destructor-dummy');
+import('components/super/i-block/test/b-super-i-block-deactivation-dummy');
+import('components/super/i-block/test/b-super-i-block-teleport-dummy');
 //#endif
 
 export * from 'core/component';
@@ -82,5 +84,83 @@ export default abstract class iBlock extends iBlockProviders {
 	 */
 	isComponent<T extends iBlock>(obj: unknown, constructor?: {new(): T} | Function): obj is T {
 		return Object.isTruly(obj) && (<Dictionary>obj).instance instanceof (constructor ?? iBlock);
+	}
+
+	/**
+	 * Handler: fixes the issue where the teleported component
+	 * and its DOM nodes were rendered before the teleport container was ready
+	 */
+	@watch({
+		path: 'r.shouldMountTeleports',
+		flush: 'post'
+	})
+
+	@hook('before:mounted')
+	protected onMountTeleports(): void {
+		const getNode = () => this.$refs[this.$resolveRef('$el')] ?? this.$el;
+
+		const {
+			$el: originalNode,
+			$async: $a
+		} = this;
+
+		const
+			node = getNode(),
+			mountedAttrs = new Set<string>(),
+			mountedAttrsGroup = {group: 'mountedAttrs'};
+
+		if (originalNode != null && node != null && originalNode !== node) {
+			// Fix the DOM element link to the component
+			originalNode.component = this;
+
+			// Fix the teleported DOM element link to the component
+			node.component = this;
+
+			Object.defineProperty(this.unsafe, '$el', {
+				configurable: true,
+				get: () => node
+			});
+
+			mountAttrs(this.$attrs);
+			this.watch('$attrs', {deep: true}, mountAttrs);
+		}
+
+		function mountAttrs(attrs: Dictionary<string>) {
+			$a.terminateWorker(mountedAttrsGroup);
+
+			if (node == null || originalNode == null) {
+				return;
+			}
+
+			Object.entries(attrs).forEach(([name, attr]) => {
+				if (attr == null) {
+					return;
+				}
+
+				if (name === 'class') {
+					attr.split(/\s+/).forEach((val) => {
+						node.classList.add(val);
+						mountedAttrs.add(`class.${val}`);
+					});
+
+				} else if (originalNode.hasAttribute(name)) {
+					node.setAttribute(name, attr);
+					mountedAttrs.add(name);
+				}
+			});
+
+			$a.worker(() => {
+				mountedAttrs.forEach((attr) => {
+					if (attr.startsWith('class.')) {
+						node.classList.remove(attr.split('.')[1]);
+
+					} else {
+						node.removeAttribute(attr);
+					}
+				});
+
+				mountedAttrs.clear();
+			}, mountedAttrsGroup);
+		}
 	}
 }