+++ src/core/component/watch/create.ts
@@ -267,6 +267,9 @@ export function createWatchFn(component: ComponentInterface): ComponentInterface
 						return res;
 					}
 				};
+
+			} else if (flush === 'post') {
+				handler = (...args) => component.$nextTick().then(() => originalHandler.call(this, ...args));
 			}
 
 			if (needImmediate) {